Obtain moving average map series

Obtain a moving average for monthly, annual or seasonal resolution data from monthly map data.

get_ma(x, type, season = NULL, size = 10, format = "table",
  use_mclapply = FALSE, mc.cores = 32)

Arguments

x
a data frame (or data table) or a list of these.

type
character, one of "monthly", "annual", or "seasonal".

season
NULL or character, one of "winter", "spring", "summer", or "autumn". Default is NULL.

size
number of years for the moving average window. Default 10.

format
return results as a single data frame with format="table" (default) or a as a list split either on unique year for seasonal and annual moving averages or on unique year and month combination for monthly moving averages.

use_mclapply
use mclapply from the parallel package (Linux). Defaults to FALSE.

mc.cores
number of CPUs for parallel processing when use_mclapply=TRUE.

Value

returns a single data frame or a list of data frames.

Details

get_ma takes a single data frame or a list of data frames. A list is useful for example if the table is very large and is pre-split into a list of smaller data frames for parallel processing on a Linux cluster with many CPU cores available. If x is a data frame rather than a data table, it will be converted to a data table.

For each data frame get_ma computes and returns the moving or rolling average, after first summarizing monthly data to seasonal or annual averages if applicable. Winter begins in December and each season is three consecutive months (DJF, MAM, JJA, SON). Whether x is a data frame or list of data frames, the output can be returned as a single data frame or list of data frames

Parallel processing is Linux-only (uses mclapply) and the default of 32 CPU cores will have to be changed if you don't have that many. It is convenient for my usage and this package version is not intended for general use. Most other users will probably not be using parallel processing at all, in which case n.cores is ignored.

Examples

data(monthlytemps)
res <- "seasonal" # annual, seasonal, monthly
season <- "winter" # winter, spring, summer, autumn
idx <- switch(season, winter=c(12,1,2), spring=3:5, summer=6:8, autumn=9:11)
if(res=="seasonal") monthlytemps <- dplyr::filter(monthlytemps, Month %in% idx)
get_ma(monthlytemps, res, season)

#> # A tibble: 80 x 4
#>          lon      lat  Year         z
#>        <dbl>    <dbl> <dbl>     <dbl>
#>  1 -176.6667 53.66633  2015 0.5263333
#>  2 -176.6667 53.66633  2016 0.4096667
#>  3 -176.6667 53.66633  2017 0.5216667
#>  4 -176.6667 53.66633  2018 0.4596667
#>  5 -176.6667 53.66633  2019 0.5473333
#>  6 -176.6667 53.66633  2020 0.4766667
#>  7 -176.6667 53.66633  2021 0.5983333
#>  8 -176.6667 53.66633  2022 0.7020000
#>  9 -176.6667 53.66633  2023 0.6650000
#> 10 -176.6667 53.66633  2024 0.7126667
#> # ... with 70 more rows

get_ma(list(monthlytemps, monthlytemps), res, season)

#> # A tibble: 160 x 4
#>          lon      lat  Year         z
#>        <dbl>    <dbl> <dbl>     <dbl>
#>  1 -176.6667 53.66633  2015 0.5263333
#>  2 -176.6667 53.66633  2016 0.4096667
#>  3 -176.6667 53.66633  2017 0.5216667
#>  4 -176.6667 53.66633  2018 0.4596667
#>  5 -176.6667 53.66633  2019 0.5473333
#>  6 -176.6667 53.66633  2020 0.4766667
#>  7 -176.6667 53.66633  2021 0.5983333
#>  8 -176.6667 53.66633  2022 0.7020000
#>  9 -176.6667 53.66633  2023 0.6650000
#> 10 -176.6667 53.66633  2024 0.7126667
#> # ... with 150 more rows

get_ma(monthlytemps, res, season, format="list")

#> $`2015`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2015 0.5263333
#> 
#> $`2016`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2016 0.4096667
#> 
#> $`2017`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2017 0.5216667
#> 
#> $`2018`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2018 0.4596667
#> 
#> $`2019`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2019 0.5473333
#> 
#> $`2020`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2020 0.4766667
#> 
#> $`2021`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2021 0.5983333
#> 
#> $`2022`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2022 0.702
#> 
#> $`2023`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2023 0.665
#> 
#> $`2024`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2024 0.7126667
#> 
#> $`2025`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2025 0.6676667
#> 
#> $`2026`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2026 0.826
#> 
#> $`2027`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2027 0.8686667
#> 
#> $`2028`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2028 0.9366667
#> 
#> $`2029`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2029 0.9626667
#> 
#> $`2030`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2030 1.038667
#> 
#> $`2031`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2031 1.067
#> 
#> $`2032`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2032 0.9796667
#> 
#> $`2033`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2033 0.9923333
#> 
#> $`2034`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2034 0.9416667
#> 
#> $`2035`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2035 1.086667
#> 
#> $`2036`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2036 0.995
#> 
#> $`2037`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2037 0.9916667
#> 
#> $`2038`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2038 0.9893333
#> 
#> $`2039`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2039 0.9363333
#> 
#> $`2040`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2040 0.886
#> 
#> $`2041`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2041 0.8516667
#> 
#> $`2042`
#> # A tibble: 1 x 4
#>         lon      lat  Year         z
#>       <dbl>    <dbl> <dbl>     <dbl>
#> 1 -176.6667 53.66633  2042 0.9603333
#> 
#> $`2043`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2043  1.01
#> 
#> $`2044`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2044 1.026
#> 
#> $`2045`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2045  0.98
#> 
#> $`2046`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2046 1.041
#> 
#> $`2047`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2047 1.182333
#> 
#> $`2048`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2048 1.183
#> 
#> $`2049`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2049 1.279
#> 
#> $`2050`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2050  1.24
#> 
#> $`2051`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2051 1.250667
#> 
#> $`2052`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2052 1.279
#> 
#> $`2053`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2053 1.432333
#> 
#> $`2054`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2054 1.544333
#> 
#> $`2055`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2055  1.52
#> 
#> $`2056`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2056 1.549333
#> 
#> $`2057`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2057 1.454667
#> 
#> $`2058`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2058 1.518333
#> 
#> $`2059`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2059 1.500333
#> 
#> $`2060`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2060 1.652333
#> 
#> $`2061`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2061 1.716667
#> 
#> $`2062`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2062  1.79
#> 
#> $`2063`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2063 1.744
#> 
#> $`2064`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2064 1.819667
#> 
#> $`2065`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2065 1.935667
#> 
#> $`2066`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2066 2.077667
#> 
#> $`2067`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2067 2.139667
#> 
#> $`2068`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2068 2.161667
#> 
#> $`2069`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2069 2.137333
#> 
#> $`2070`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2070 2.094667
#> 
#> $`2071`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2071 2.011333
#> 
#> $`2072`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2072 1.983333
#> 
#> $`2073`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2073 2.010333
#> 
#> $`2074`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2074 1.894333
#> 
#> $`2075`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2075 1.767
#> 
#> $`2076`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2076 1.649
#> 
#> $`2077`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2077 1.600667
#> 
#> $`2078`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2078 1.574333
#> 
#> $`2079`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2079 1.712333
#> 
#> $`2080`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2080 1.772667
#> 
#> $`2081`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2081 1.834667
#> 
#> $`2082`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2082 1.798333
#> 
#> $`2083`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2083 1.793667
#> 
#> $`2084`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2084 1.840667
#> 
#> $`2085`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2085 1.949
#> 
#> $`2086`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2086 1.977333
#> 
#> $`2087`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2087 1.951
#> 
#> $`2088`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2088 1.966
#> 
#> $`2089`
#> # A tibble: 1 x 4
#>         lon      lat  Year        z
#>       <dbl>    <dbl> <dbl>    <dbl>
#> 1 -176.6667 53.66633  2089 1.824667
#> 
#> $`2090`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2090 1.763
#> 
#> $`2091`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2091 1.774
#> 
#> $`2092`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2092  1.84
#> 
#> $`2093`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2093 1.787
#> 
#> $`2094`
#> # A tibble: 1 x 4
#>         lon      lat  Year     z
#>       <dbl>    <dbl> <dbl> <dbl>
#> 1 -176.6667 53.66633  2094 1.815
#> 
# NOT RUN {
get_ma(list(monthlytemps, monthlytemps), res, season, use_mclapply=T, mc.cores=2)
# }