Identidy visible points on an arbitrary global hemishpere view.

Given a global hemispheric field of view defined by a single latitudinal and longitudinal centroid focal point, project geographic points onto the hemishpere.

project_to_hemisphere(lon, lat, lon0, lat0)

Arguments

lon
vector of longitudes.

lat
vector of latitudes.

lon0
longitude of focus coordinates.

lat0
latitude of focus coordinates.

Value

returns a data frame.

Details

project_to_hemisphere identifies whether each pair of coordinates in the lat and lon vectors is in a field of view defined by a centroid focal point (lat0, lon0) and returns a data frame containing the original coordinates and a column indicating if the coordinates are in the field of view (TRUE or FALSE).

Examples

lon <- seq(-180, 180, length.out=60)
lat <- rep(seq(-90, 90, length.out=30), 2)
project_to_hemisphere(lon, lat , 0, 0)

#> # A tibble: 60 x 3
#>          lon       lat inview
#>        <dbl>     <dbl>  <lgl>
#>  1 -180.0000 -90.00000  FALSE
#>  2 -173.8983 -83.79310  FALSE
#>  3 -167.7966 -77.58621  FALSE
#>  4 -161.6949 -71.37931  FALSE
#>  5 -155.5932 -65.17241  FALSE
#>  6 -149.4915 -58.96552  FALSE
#>  7 -143.3898 -52.75862  FALSE
#>  8 -137.2881 -46.55172  FALSE
#>  9 -131.1864 -40.34483  FALSE
#> 10 -125.0847 -34.13793  FALSE
#> # ... with 50 more rows